refactored validator structure
