+++ src/pages/exercise/[id]/edit.tsx
@@ -3,13 +3,19 @@ import { authorize } from "@/utils/authorize";
 
 export { authorize as getServerSideProps };
 
+const EditExerciseForm = () => {
+  const;
+  return null;
+};
+
 export default function EditExercisePage() {
   return (
     <>
       <Header app />
 
       <main className='container mx-auto p-4 grid place-content-center min-h-screen'>
-        <h1>Feature coming soon!</h1>
+        <h1>Edit Exercise</h1>
+        <EditExerciseForm />
       </main>
     </>
   );
