+++ src/utils/validators/exercise.ts
@@ -1,0 +1,27 @@
+import { z } from "zod";
+
+export const exerciseValidator = z.object({
+  name: z.string().min(1, "You must provide a name"),
+  currentWeight: z
+    .string()
+    .transform(data => Number(data))
+    .or(z.number())
+    .optional(),
+  targetWeight: z
+    .string()
+    .transform(data => Number(data))
+    .or(z.number())
+    .optional(),
+  currentDistance: z
+    .string()
+    .transform(data => Number(data))
+    .or(z.number())
+    .optional(),
+  targetDistance: z
+    .string()
+    .transform(data => Number(data))
+    .or(z.number())
+    .optional()
+});
+
+export type ExerciseValidator = z.infer<typeof exerciseValidator>;
